2.1

+++ lib/data/sync/timetable.dart
@@ -1,44 +1,52 @@
 import 'dart:convert';
-
+import 'package:filcnaplo/ui/pages/planner/timetable/builder.dart';
 import 'package:filcnaplo/data/context/app.dart';
 import 'package:filcnaplo/data/models/lesson.dart';
-//import 'package:filcnaplo/data/models/dummy.dart';
+import 'package:filcnaplo/ui/pages/planner/timetable/week.dart';
 
 class TimetableSync {
-  List<Lesson> data = [];
+  List<Lesson> lessons = [];
   DateTime from;
   DateTime to;
 
   Future<bool> sync() async {
+    Week currentWeek =
+        TimetableBuilder().getWeek(TimetableBuilder().getCurrentWeek());
+    bool updatingCurrent =
+        (currentWeek.start == from) && (currentWeek.end == to);
+
     if (!app.debugUser) {
-      List<Lesson> lessons;
-      lessons = await app.user.kreta.getLessons(from, to);
+      List<Lesson> _lessons;
+      _lessons = await app.user.kreta.getLessons(from, to);
 
-      if (lessons == null) {
+      if (_lessons == null) {
         await app.user.kreta.refreshLogin();
-        lessons = await app.user.kreta.getLessons(from, to);
+        _lessons = await app.user.kreta.getLessons(from, to);
       }
 
-      if (lessons != null) {
-        data = lessons;
+      if (_lessons != null) {
+        lessons = _lessons;
 
-        await app.user.storage.delete("kreta_lessons");
+        //Only do DB calls if we're working with the current week.
+        if (updatingCurrent) {
+          await app.user.storage.delete("kreta_lessons");
 
-        await Future.forEach(lessons, (lesson) async {
-          if (lesson.json != null &&
-              from.isBefore(DateTime.now()) &&
-              to.isAfter(DateTime.now())) {
-            await app.user.storage.insert("kreta_lessons", {
-              "json": jsonEncode(lesson.json),
-            });
-          }
-        });
+          await Future.forEach(_lessons, (lesson) async {
+            if (lesson.json != null &&
+                from.isBefore(DateTime.now()) &&
+                to.isAfter(DateTime.now())) {
+              await app.user.storage.insert("kreta_lessons", {
+                "json": jsonEncode(lesson.json),
+              });
+            }
+          });
+        }
       } else {
-        data = [];
-        await app.user.storage.delete("kreta_lessons");
+        lessons = [];
+        if (updatingCurrent) await app.user.storage.delete("kreta_lessons");
       }
 
-      return lessons != null;
+      return _lessons != null;
     } else {
       //data[0] = Dummy.lessons;
       return true;
@@ -46,6 +54,6 @@ class TimetableSync {
   }
 
   delete() {
-    data = [];
+    lessons = [];
   }
 }
